move validate_forecast_values
