+++ sfa_api/forecasts.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, request, jsonify, make_response, url_for
 from flask.views import MethodView
 from marshmallow import ValidationError
-import pandas as pd
 
 
 from sfa_api import spec
@@ -18,44 +17,8 @@ from sfa_api.utils.errors import BadAPIRequest
 from sfa_api.utils.storage import get_storage
 from sfa_api.utils.request_handling import (validate_parsable_values,
                                             validate_start_end,
-                                            validate_index_period)
-
-
-def validate_forecast_values(forecast_df):
-    """Validates that posted values are parseable and of the expectedtypes.
-
-    Parameters
-    ----------
-    forecast_df: Pandas DataFrame
-
-    Raises
-    ------
-    BadAPIRequestError
-        If an expected field is missing or contains an entry of incorrect
-        type.
-    """
-    errors = {}
-    try:
-        forecast_df['value'] = pd.to_numeric(forecast_df['value'],
-                                             downcast='float')
-    except ValueError:
-        error = ('Invalid item in "value" field. Ensure that all values '
-                 'are integers, floats, empty, NaN, or NULL.')
-        errors.update({'value': [error]})
-    except KeyError:
-        errors.update({'value': ['Missing "value" field.']})
-    try:
-        forecast_df['timestamp'] = pd.to_datetime(
-            forecast_df['timestamp'],
-            utc=True)
-    except ValueError:
-        error = ('Invalid item in "timestamp" field. Ensure that '
-                 'timestamps are ISO8601 compliant')
-        errors.update({'timestamp': [error]})
-    except KeyError:
-        errors.update({'timestamp': ['Missing "timestamp" field.']})
-    if errors:
-        raise BadAPIRequest(errors)
+                                            validate_index_period,
+                                            validate_forecast_values)
 
 
 class AllForecastsView(MethodView):
